+++ test/main.test.ts
@@ -1,14 +1,15 @@
 import { when } from "jest-when";
 
+import * as github from "./mocks/@actions/github.mock";
 import * as core from "./mocks/@actions/core.mock";
-import * as fs from "./mocks/file-system.mock";
+import * as fileSystems from "./mocks/file-systems.mock";
 import * as inputs from "./mocks/inputs.mock";
 import * as optimize from "./mocks/optimize.mock";
 import * as outputs from "./mocks/outputs.mock";
 import * as parser from "./mocks/parser.mock";
 import * as svgo from "./mocks/svgo.mock";
 
-jest.mock("../src/file-system", () => fs);
+jest.mock("../src/file-systems", () => fileSystems);
 jest.mock("../src/inputs", () => inputs);
 jest.mock("../src/optimize", () => optimize);
 jest.mock("../src/outputs", () => outputs);
@@ -22,7 +23,16 @@ import {
 } from "../src/constants";
 import main from "../src/main";
 
-const DEFAULT_CONTEXT = { eventName: EVENT_PUSH };
+const DEFAULT_CONTEXT = {
+  eventName: EVENT_PUSH,
+  payload: { },
+  repo: {
+    owner: "pikachu",
+    repo: "pokédex",
+  },
+};
+
+const DEFAULT_CLIENT = github.getOctokit();
 
 beforeEach(() => {
   core.debug.mockClear();
@@ -33,28 +43,30 @@ beforeEach(() => {
 
 describe("run & optimize", () => {
   test.each(SUPPORTED_EVENTS)("%s event", async (eventName) => {
-    const context = { eventName };
+    const context = Object.assign({ }, DEFAULT_CONTEXT, { eventName });
 
-    await main(core, context);
+    await main(core, DEFAULT_CLIENT, context);
     expect(optimize.optimize).toHaveBeenCalled();
   });
 
   test.each(SUPPORTED_EVENTS)("%s event error", async (eventName) => {
-    const context = { eventName };
+    const context = Object.assign({}, DEFAULT_CONTEXT, { eventName });
 
     optimize.optimize.mockImplementationOnce(() => {
       throw new Error("Something went wrong");
     });
 
-    await main(core, context);
+    await main(core, DEFAULT_CLIENT, context);
     expect(optimize.optimize).toHaveBeenCalledTimes(1);
     expect(core.setFailed).toHaveBeenCalledTimes(1);
   });
 
   test("unknown event", async () => {
-    const context = { eventName: "UnKnOwN eVeNt" };
+    const context = Object.assign({}, DEFAULT_CONTEXT, {
+      eventName: "UnKnOwN eVeNt",
+    });
 
-    await main(core, context);
+    await main(core, DEFAULT_CLIENT, context);
     expect(optimize.optimize).not.toHaveBeenCalled();
     expect(core.setFailed).toHaveBeenCalledTimes(1);
   });
@@ -62,13 +74,13 @@ describe("run & optimize", () => {
 
 describe("configuration", () => {
   test.each(SUPPORTED_EVENTS)("dry mode enabled (%s)", async (eventName) => {
-    const context = { eventName };
+    const context = Object.assign({}, DEFAULT_CONTEXT, { eventName });
 
     inputs.ActionConfig.mockImplementationOnce(() => {
       return { svgoOptionsPath: "svgo.config.js", isDryRun: true };
     });
 
-    await main(core, context);
+    await main(core, DEFAULT_CLIENT, context);
     expect(core.info).toHaveBeenCalledWith(expect.stringContaining("Dry mode enabled"));
   });
 
@@ -80,7 +92,7 @@ describe("configuration", () => {
       return { svgoOptionsPath: "svgo.config.js", svgoVersion: svgoVersion };
     });
 
-    await main(core, DEFAULT_CONTEXT);
+    await main(core, DEFAULT_CLIENT, DEFAULT_CONTEXT);
 
     expect(svgo.SVGOptimizer).toHaveBeenCalledWith(svgoVersion, expect.anything());
     expect(core.info).toHaveBeenCalledWith(expect.stringMatching(`SVGO.*${svgoVersion}`));
@@ -89,11 +101,11 @@ describe("configuration", () => {
 
 describe("SVGO configuration file", () => {
   test.each(SUPPORTED_EVENTS)("use a JavaScript SVGO options file in the repository (%s)", async (eventName) => {
-    fs.readFile.mockClear();
+    fileSystems.readFile.mockClear();
     parser.parseJavaScript.mockClear();
     svgo.SVGOptimizer.mockClear();
 
-    const context = { eventName };
+    const context = Object.assign({}, DEFAULT_CONTEXT, { eventName });
 
     const svgoOptionsFilePath = "svgo.config.js";
     const svgoOptionsFileContent = "module.exports = { }";
@@ -102,26 +114,26 @@ describe("SVGO configuration file", () => {
     inputs.ActionConfig.mockImplementationOnce(() => {
       return { svgoOptionsPath: svgoOptionsFilePath, svgoVersion: 2 };
     });
-    when(fs.readFile)
+    when(fileSystems.readFile)
       .calledWith(svgoOptionsFilePath)
       .mockReturnValueOnce(svgoOptionsFileContent);
     when(parser.parseJavaScript)
       .calledWith(svgoOptionsFileContent)
       .mockReturnValueOnce(svgoOptions);
 
-    await main(core, context);
+    await main(core, DEFAULT_CLIENT, context);
 
-    expect(fs.readFile).toHaveBeenCalledWith(svgoOptionsFilePath);
+    expect(fileSystems.readFile).toHaveBeenCalledWith(svgoOptionsFilePath);
     expect(parser.parseJavaScript).toHaveBeenCalledWith(svgoOptionsFileContent);
     expect(svgo.SVGOptimizer).toHaveBeenCalledWith(2, svgoOptions);
   });
 
   test.each(SUPPORTED_EVENTS)("use a YAML SVGO options file in the repository (%s)", async (eventName) => {
-    fs.readFile.mockClear();
+    fileSystems.readFile.mockClear();
     parser.parseYaml.mockClear();
     svgo.SVGOptimizer.mockClear();
 
-    const context = { eventName };
+    const context = Object.assign({}, DEFAULT_CONTEXT, { eventName });
 
     const svgoOptionsFilePath = ".svgo.yml";
     const svgoOptionsFileContent = "multipass: true\nplugins:\n- removeDoctype";
@@ -130,16 +142,16 @@ describe("SVGO configuration file", () => {
     inputs.ActionConfig.mockImplementationOnce(() => {
       return { svgoOptionsPath: svgoOptionsFilePath, svgoVersion: 1 };
     });
-    when(fs.readFile)
+    when(fileSystems.readFile)
       .calledWith(svgoOptionsFilePath)
       .mockReturnValueOnce(svgoOptionsFileContent);
     when(parser.parseYaml)
       .calledWith(svgoOptionsFileContent)
       .mockReturnValueOnce(svgoOptions);
 
-    await main(core, context);
+    await main(core, DEFAULT_CLIENT, context);
 
-    expect(fs.readFile).toHaveBeenCalledWith(svgoOptionsFilePath);
+    expect(fileSystems.readFile).toHaveBeenCalledWith(svgoOptionsFilePath);
     expect(parser.parseYaml).toHaveBeenCalledWith(svgoOptionsFileContent);
     expect(svgo.SVGOptimizer).toHaveBeenCalledWith(1, svgoOptions);
   });
@@ -153,11 +165,11 @@ describe("SVGO configuration file", () => {
       return { svgoOptionsPath: DEFAULT_SVGO_OPTIONS, svgoVersion: 2 };
     });
 
-    when(fs.readFile)
+    when(fileSystems.readFile)
       .calledWith(DEFAULT_SVGO_OPTIONS)
       .mockRejectedValueOnce(new Error("Not found"));
 
-    await main(core, DEFAULT_CONTEXT);
+    await main(core, DEFAULT_CLIENT, DEFAULT_CONTEXT);
 
     expect(core.setFailed).not.toHaveBeenCalled();
     expect(core.warning).not.toHaveBeenCalled();
@@ -176,11 +188,11 @@ describe("SVGO configuration file", () => {
       return { svgoOptionsPath: filePath, svgoVersion: version };
     });
 
-    when(fs.readFile)
+    when(fileSystems.readFile)
       .calledWith(filePath)
       .mockRejectedValueOnce(new Error("Not found"));
 
-    await main(core, DEFAULT_CONTEXT);
+    await main(core, DEFAULT_CLIENT, DEFAULT_CONTEXT);
 
     expect(core.setFailed).not.toHaveBeenCalled();
     expect(core.warning).toHaveBeenCalledWith(
@@ -203,7 +215,7 @@ describe("SVGO configuration file", () => {
       return { svgoOptionsPath: filePath, svgoVersion: version };
     });
 
-    when(fs.readFile)
+    when(fileSystems.readFile)
       .calledWith(filePath)
       .mockResolvedValueOnce(invalidContent);
     when(parser.parseJavaScript)
@@ -213,7 +225,7 @@ describe("SVGO configuration file", () => {
       .calledWith(invalidContent)
       .mockImplementation(() => { throw new Error("Not found"); });
 
-    await main(core, DEFAULT_CONTEXT);
+    await main(core, DEFAULT_CLIENT, DEFAULT_CONTEXT);
 
     expect(core.setFailed).not.toHaveBeenCalled();
     expect(core.warning).toHaveBeenCalledWith(